self_debug/pyproject.toml
[project]
name = "self-debug"
version = "0.1.0"
description = "SDFeedback: Code migration with LLMs using self-debugging feedback"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    # AWS
    "boto3>=1.34.4",

    # Git
    "gitpython>=3.1.43",

    # Java parsing
    "javalang>=0.13.0",

    # Core utilities
    "numpy>=1.26.4",
    "protobuf>=3.20.3",
    "pydantic>=2.6.4",
    "pytz>=2024.1",
    "packaging>=25.0",
    "requests>=2.32.3",

    # Spark for batch processing
    "pyspark>=3.5.1",

    # HuggingFace datasets
    "datasets>=2.14.0",
    "huggingface-hub>=0.16.0",

    # Evaluation from migration-bench
    "migrationbench@git+https://github.com/amazon-science/MigrationBench.git",
]

[project.optional-dependencies]
dev = [
    "parameterized>=0.8.1",
    "pylint>=2.14.5",
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/self_debug"]

[tool.uv]
index-url = "https://pypi.org/simple"
